backend-v6

+++ classManagement_BE/src/main/java/ptit/d20/do_an/class_management/service/ClassScheduleService.java
@@ -153,6 +153,10 @@ public class ClassScheduleService {
         schedules.stream()
                 .filter(schedule -> schedule.getId().equals(scheduleId)).findFirst()
                 .orElseThrow(() -> new ResourceNotFoundException("Not found schedule"));
+
+        boolean hasRelevantAttendance = classAttendanceRepository.existsByClassScheduleIdAndIsAttended(scheduleId, true);
+        if (hasRelevantAttendance) throw new BusinessException("Error");
+
         try {
             classAttendanceRepository.deleteAllByClassScheduleId(scheduleId);
             classScheduleRepository.deleteById(scheduleId);
